pull entity generation functions into utils
